requirements.txt
streamlit==1.31.1
beautifulsoup4==4.12.3
requests==2.31.0
nltk==3.8.1
transformers==4.37.2
torch==2.2.0
pandas==2.2.0
numpy==1.26.3
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
gTTS==2.5.1
textblob==0.17.1
newspaper3k==0.2.8
plotly==5.18.0
lxml[html_clean]